@@ -1,23 +1,22 @@
 ﻿using System;
 using Nest;
 
-namespace Bing.Elasticsearch.Tests.Models
+namespace Bing.Elasticsearch.Tests.Models;
+
+[ElasticsearchType(RelationName = "test_student", IdProperty = "StudentId")]
+public class StudentSample
 {
-    [ElasticsearchType(RelationName = "test_student", IdProperty = "StudentId")]
-    public class StudentSample
-    {
-        public long StudentId { get; set; }
+    public long StudentId { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public int Age { get; set; }
+    public int Age { get; set; }
 
-        public string Address { get; set; }
+    public string Address { get; set; }
 
-        public DateTime BirthDay { get; set; }
+    public DateTime BirthDay { get; set; }
 
-        public bool IsValid { get; set; }
+    public bool IsValid { get; set; }
 
-        public Guid UserId { get; set; }
-    }
+    public Guid UserId { get; set; }
 }
